moved to new package -> viewmodels
